Bug #3485096 & #3485453
I found a proble with importers/nagios/NagiosServiceImporter.php, approx line 506, with block "re-merge the check_commands" and the use of the variable by reference. If we print the value $values variable ​​before and after this block, we can see how lost the last value:
We see that when you import FileSystemCheck service from example01.cfg (lines 229-236).

I found a problem with importers/nagios/NagiosHostImporter.php, approx line 272, with block "re-merge the check_commands" and the use of the variable by reference: This problem is similar to BUG ID: 3485096, but with Host importer, and does not import contact_groups.