Fix bug for empty contact or contactgroup list
Fix two bugs where, when no contacts or contactgroups were defined for a service, the exporter would output a line that said "contacts " or "contactgroups " with no other data, causing an error. Fix was to only output these lines when there is data to put on the lines. (such as "contacts johnsmith" or "contactgroups admins")